# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MKWW MCP Server API
  description: |
    # MKWW MCP Server API
    
    This is the API documentation for the MKWW MCP (Master Control Program) Server.
    The MCP Server acts as the central orchestration hub for the MKWorldWide ecosystem.
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the `Authorization` header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## WebSocket API
    
    The MCP Server provides a WebSocket endpoint for real-time communication. Connect to:
    
    ```
    ws://<host>:<port>/ws
    ```
    
    ### WebSocket Message Format
    
    All WebSocket messages are JSON objects with the following structure:
    
    ```json
    {
      "type": "message-type",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: MKWW Support
    email: support@mkww.io
  license:
    name: Proprietary
    url: https://mkww.io/terms
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.mkww.io/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        priority:
          type: integer
          minimum: 1
          maximum: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        result:
          type: object
          additionalProperties: true
        error:
          type: string
          nullable: true
...
paths:
...
  /tasks:
    get:
      summary: List tasks
      description: Retrieve a list of tasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, in_progress, completed, failed]
          description: Filter tasks by status
        - in: query
          name: type
          schema:
            type: string
          description: Filter tasks by type
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 1000
          description: Maximum number of tasks to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of tasks to skip
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /tasks/{taskId}:
    get:
      summary: Get a task
      description: Retrieve a specific task by ID
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the task to retrieve
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

# WebSocket API
...